pyproject.toml
[project]
name = "eag-v2-s10-custom"
version = "0.1.0"
description = "Custom Multi-Agent System for EAG V2 Session 10"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.3.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.2.0",
    "pyyaml>=6.0.0",
    "colorama>=0.4.6",
    "termcolor>=2.4.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "faiss-cpu>=1.7.4",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "tqdm>=4.66.5",
    "httpx>=0.27.0",
    "markitdown>=0.0.1",
    "trafilatura>=1.9.0",
    "pymupdf4llm>=0.0.1",
    "duckduckgo-search>=8.1.1",
    "tavily-python>=0.7.13",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0"
]
